+++ db/area.go
@@ -71,7 +71,7 @@ func (n Areas) GetByName(name string) (area *Area, err error) {
 }
 
 // Search ...
-func (n *Areas) Search(query string, limit, offset int) (list []*Area, total int64, err error) {
+func (n *Areas) Search(query string, limit, offset int64) (list []*Area, total int64, err error) {
 
 	q := n.Db.Model(&Area{}).
 		Where("name LIKE ?", "%"+query+"%")
